refactor logtail related code (#5399)
Refactor logtail related code

Approved by: @yingfeng
