+++ pkg/vm/engine/tae/iface/txnif/types.go
@@ -233,16 +233,10 @@ type TxnStore interface {
 	IsReadonly() bool
 	IncreateWriteCnt() int
 
-	HasTableDataChanges(tableID uint64) bool
+	HasAnyTableDataChanges() bool
+	HasTableDataChanges(id uint64) bool
 	HasCatalogChanges() bool
-	GetTableDirtyPoints(tableID uint64) DirtySet
-}
-
-type DirtySet = map[DirtyPoint]struct{}
-
-// not use common id to save space, less hash cost
-type DirtyPoint struct {
-	SegID, BlkID uint64
+	GetDirtyTableByID(id uint64) *common.TableTree
 }
 
 type TxnEntryType int16
