+++ pkg/vm/engine/tae/txn/txnimpl/localseg.go
@@ -178,7 +178,7 @@ func (seg *localSegment) prepareApplyNode(node InsertNode) (err error) {
 		}
 		id := appender.GetID()
 		seg.table.store.warChecker.ReadBlock(seg.table.entry.GetDB().ID, id)
-		seg.table.store.dirtyMemo.recordBlk(*id)
+		seg.table.store.dirtyMemo.recordBlk(seg.table.entry.GetDB().ID, id)
 		seg.appends = append(seg.appends, ctx)
 		logutil.Debugf("%s: toAppend %d, appended %d, blks=%d", id.String(), toAppend, appended, len(seg.appends))
 		appended += toAppend
