+++ pkg/vm/engine/tae/txn/txnimpl/store.go
@@ -32,48 +32,29 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/wal"
 )
 
-type dirtyPoint = txnif.DirtyPoint
-type dirtySet = txnif.DirtySet
-
 // dirtyMemo intercepts txn to record changed segments and blocks, or catalog
 //
 // no locks to protect dirtyMemo because it is expected to be quried by other goroutines
 // until the txn enqueued, and after that, no one will change it
 type dirtyMemo struct {
 	// tableChanges records modified segments and blocks in current txn
-	tableChanges map[uint64]dirtySet
+	tableChanges *common.Tree
 	// catalogChanged indicates whether create/drop db/table
 	catalogChanged bool
 }
 
 func newDirtyMemo() *dirtyMemo {
 	return &dirtyMemo{
-		tableChanges: make(map[uint64]dirtySet, 0),
+		tableChanges: common.NewTree(),
 	}
 }
 
-func (m *dirtyMemo) recordBlk(id common.ID) {
-	point := dirtyPoint{
-		SegID: id.SegmentID,
-		BlkID: id.BlockID,
-	}
-	m.recordDirty(id.TableID, point)
+func (m *dirtyMemo) recordBlk(dbID uint64, id *common.ID) {
+	m.tableChanges.AddBlock(dbID, id.TableID, id.SegmentID, id.BlockID)
 }
 
-func (m *dirtyMemo) recordSeg(tid, sid uint64) {
-	point := dirtyPoint{
-		SegID: sid,
-	}
-	m.recordDirty(tid, point)
-}
-
-func (m *dirtyMemo) recordDirty(tid uint64, point dirtyPoint) {
-	dirties, exist := m.tableChanges[tid]
-	if !exist {
-		dirties = make(map[dirtyPoint]struct{}, 0)
-		m.tableChanges[tid] = dirties
-	}
-	dirties[point] = struct{}{}
+func (m *dirtyMemo) recordSeg(dbID, tableID, segID uint64) {
+	m.tableChanges.AddSegment(dbID, tableID, segID)
 }
 
 func (m *dirtyMemo) recordCatalogChange() {
@@ -551,15 +532,17 @@ func (store *txnStore) PrepareRollback() error {
 
 func (store *txnStore) GetLSN() uint64 { return store.cmdMgr.lsn }
 
-func (store *txnStore) HasTableDataChanges(tableID uint64) bool {
-	_, changed := store.dirtyMemo.tableChanges[tableID]
-	return changed
+func (store *txnStore) HasAnyTableDataChanges() bool {
+	return store.dirtyMemo.tableChanges.TableCount() > 0
 }
 
-// GetTableDirtyPoints returns touched segments and blocks in the txn.
-func (store *txnStore) GetTableDirtyPoints(tableID uint64) dirtySet {
-	dirtiesMap := store.dirtyMemo.tableChanges[tableID]
-	return dirtiesMap
+func (store *txnStore) HasTableDataChanges(id uint64) bool {
+	return store.dirtyMemo.tableChanges.HasTable(id)
+}
+
+// GetDirtyTable returns touched segments and blocks in the txn.
+func (store *txnStore) GetDirtyTableByID(id uint64) *common.TableTree {
+	return store.dirtyMemo.tableChanges.GetTable(id)
 }
 
 func (store *txnStore) HasCatalogChanges() bool {
